+++ passes/R011/testdata/src/a/outside_package.go
@@ -1,0 +1,12 @@
+package a
+
+import (
+	"a/schema"
+)
+
+func foutside() {
+	_ = schema.Resource{
+		MigrateState:  migrateStateFunc,
+		SchemaVersion: 1,
+	}
+}