+++ passes/R011/testdata/src/a/schema/schema.go
@@ -1,0 +1,10 @@
+package schema
+
+import (
+	"github.com/hashicorp/terraform-plugin-sdk/terraform"
+)
+
+type Resource struct {
+	MigrateState  func(int, *terraform.InstanceState, interface{}) (*terraform.InstanceState, error)
+	SchemaVersion int
+}