+++ passes/checks.go
@@ -18,6 +18,7 @@ import (
 	"github.com/bflad/tfproviderlint/passes/R008"
 	"github.com/bflad/tfproviderlint/passes/R009"
 	"github.com/bflad/tfproviderlint/passes/R010"
+	"github.com/bflad/tfproviderlint/passes/R011"
 	"github.com/bflad/tfproviderlint/passes/S001"
 	"github.com/bflad/tfproviderlint/passes/S002"
 	"github.com/bflad/tfproviderlint/passes/S003"
@@ -83,6 +84,7 @@ var AllChecks = []*analysis.Analyzer{
 	R008.Analyzer,
 	R009.Analyzer,
 	R010.Analyzer,
+	R011.Analyzer,
 	S001.Analyzer,
 	S002.Analyzer,
 	S003.Analyzer,