+++ CHANGELOG.md
@@ -6,6 +6,7 @@ FEATURES
 * **New Check:** `R008`: check for deprecated `(schema.ResourceData).SetPartial` receiver method usage
 * **New Check:** `R009`: check for Go panic usage
 * **New Check:** `R010`: check for `(schema.ResourceData).GetChange` assignment which should use `(schema.ResourceData).Get`
+* **New Check:** `R011`: check for `Resource` that configure `MigrateState`
 
 # v0.9.0